refactor(common): transform test only methods on entities into extension traits
This allow a clean split between the production and test code even when
using thoses methods as to do so you will now need to import a trait
located in the `mithril_common::test` module.

one caveat: in order to transform `Blockrange::new` into a extension
trait method, its `inner_range` property had to be made `pub(crate)`.
